Moving ClusterJet to common directory
